+++ spec/03-plugins/03-http-log/01-log_spec.lua
@@ -6,6 +6,8 @@ for _, strategy in helpers.each_strategy() do
   describe("Plugin: http-log (log) [#" .. strategy .. "]", function()
     local proxy_client
     local proxy_client_grpc, proxy_client_grpcs
+    local vault_env_name = "HTTP_LOG_KEY2"
+    local vault_env_value = "the secret"
 
     lazy_setup(function()
       local bp = helpers.get_db_utils(strategy, {
@@ -214,6 +216,34 @@ for _, strategy in helpers.each_strategy() do
         }
       }
 
+      local service10 = bp.services:insert{
+        protocol = "http",
+        host     = helpers.mock_upstream_host,
+        port     = helpers.mock_upstream_port,
+      }
+
+      local route10 = bp.routes:insert {
+        hosts   = { "vault_headers_logging.test" },
+        service = service10
+      }
+
+      bp.plugins:insert {
+        route = { id = route10.id },
+        name     = "http-log",
+        config   = {
+          http_endpoint = "http://" .. helpers.mock_upstream_host
+            .. ":"
+            .. helpers.mock_upstream_port
+            .. "/post_log/vault_header",
+          headers = {
+            key1 = "value1",
+            key2 = "{vault://env/http-log-key2}"
+          }
+        }
+      }
+
+      helpers.setenv(vault_env_name, vault_env_value)
+
       assert(helpers.start_kong({
         database = strategy,
         nginx_conf = "spec/fixtures/custom_nginx.template",
@@ -224,6 +254,7 @@ for _, strategy in helpers.each_strategy() do
     end)
 
     lazy_teardown(function()
+      helpers.unsetenv(vault_env_name)
       helpers.stop_kong()
     end)
 
@@ -456,6 +487,37 @@ for _, strategy in helpers.each_strategy() do
           if ok == 2 then
             return true
           end
+        end
+      end, 10)
+    end)
+
+    it("should dereference config.headers value", function()
+      local res = assert(proxy_client:send({
+        method = "GET",
+        path = "/status/200",
+        headers = {
+          ["Host"] = "vault_headers_logging.test"
+        }
+      }))
+      assert.res_status(200, res)
+
+      helpers.wait_until(function()
+        local client = assert(helpers.http_client(helpers.mock_upstream_host,
+          helpers.mock_upstream_port))
+        local res = assert(client:send {
+          method  = "GET",
+          path    = "/read_log/vault_header",
+          headers = {
+            Accept = "application/json"
+          }
+        })
+        local raw = assert.res_status(200, res)
+        local body = cjson.decode(raw)
+
+        if #body.entries == 1 then
+          assert.same("value1", body.entries[1].log_req_headers.key1)
+          assert.same(vault_env_value, body.entries[1].log_req_headers.key2)
+          return true
         end
       end, 10)
     end)